+++ packages/eas-cli/scripts/prepack.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+set -eo pipefail
+
+rm -rf build
+yarn build
+
+if [ -z "$CLI_SIZE_CHECK" ]; then
+  yarn oclif manifest
+fi