+++ packages/eas-cli/src/commands/webhook/update.ts
@@ -1,4 +1,4 @@
-import { flags } from '@oclif/command';
+import { Flags } from '@oclif/core';
 
 import EasCommand from '../../commandUtils/EasCommand';
 import { WebhookType } from '../../graphql/generated';
@@ -12,25 +12,25 @@ export default class WebhookUpdate extends EasCommand {
   static description = 'Create a webhook on the current project.';
 
   static flags = {
-    id: flags.string({
+    id: Flags.string({
       description: 'Webhook ID',
       required: true,
     }),
-    event: flags.enum({
+    event: Flags.enum({
       description: 'Event type that triggers the webhook',
       options: [WebhookType.Build, WebhookType.Submit],
     }),
-    url: flags.string({
+    url: Flags.string({
       description: 'Webhook URL',
     }),
-    secret: flags.string({
+    secret: Flags.string({
       description:
         "Secret used to create a hash signature of the request payload, provided in the 'Expo-Signature' header.",
     }),
   };
 
   async runAsync(): Promise<void> {
-    const { flags } = this.parse(WebhookUpdate);
+    const { flags } = await this.parse(WebhookUpdate);
 
     const webhookId = flags.id;