+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This is the log of notable changes to EAS CLI and related packages.
 
 ### 🎉 New features
 
-- Add eas channel:create. ([#846](https://github.com/expo/eas-cli/pull/846) by [@jkhales](https://github.com/jkhales))
+- Add `eas channel:create`. ([#846](https://github.com/expo/eas-cli/pull/846) by [@jkhales](https://github.com/jkhales))
 
 ### 🐛 Bug fixes